added templates for user, added add/delete/show user/s

+++ src/AppBundle/Controller/UserController.php
@@ -57,7 +57,7 @@ class UserController extends Controller
             $email = $data["email"];
             //check if other user exists with new email
             if(false){
-                $error = "user with this email already exists";
+                $error = "Ein Benutzer mit dieser E-Mail Addresse existiert bereits";
             }
             $user->setEmail($email);
             $user->setUsername($email);
@@ -67,7 +67,7 @@ class UserController extends Controller
 
             //password can be max 4096 characters long
             if (strlen($plainPassword) > 4096 || strlen($plainPassword) < 1) {
-                $error = "Invalid password length";
+                $error = "Das Passwort hat eine ungültige Länge. Es muss zwischen 1 und 4069 zeichen enthalten";
             }
             //create hashed password
             $password = $this->get('security.password_encoder')
@@ -80,6 +80,7 @@ class UserController extends Controller
             //render template with new user
             return $this->render('user/show.html.twig', [
                 'user' => $user,
+                'message' => "Benutzer wurde restellt",
                 'base_dir' => realpath($this->getParameter('kernel.project_dir')).DIRECTORY_SEPARATOR,
             ]);
         }
@@ -101,22 +102,28 @@ class UserController extends Controller
             $new_user = new User();
             //get POST data
             $data = $request->request->all();
+            $error = false;
+            $name = $data["name"];
+            if($name === null || $name === ""){
+                $error = "Bitte geben sie einen Namen an.";
+            }
             $email = $data["email"];
             if($email === null || $email === ""){
-                $error = "please specify email";
+                $error = "Bitte geben sie eine E-Mail Addresse an.";
             }
+            //TODO: check if email is already taken
             //check if user with this email already exists
             $em = $this->getDoctrine()->getManager();
             $existing_user = $em->getRepository('AppBundle:User')->findOneByEmail($email);
             if($existing_user){
-                $error = "email already taken";
+                $error = "Ein Benutzer mit dieser E-Mail Addresse existiert bereits.";
             }
             $new_user->setEmail($email);
             $new_user->setUsername($email);
             $new_user->setName($data["name"]);
             $role = $data["role"];
             if(!$this->roleExists($role)){
-                $error = "invalid role";
+                $error = "Die ausgewählte Rolle ist ungültig";
             }
             $new_user->setRoles(array($role));
 
@@ -124,7 +131,7 @@ class UserController extends Controller
 
             //password can be max 4096 characters long
             if (strlen($plainPassword) > 4096 || strlen($plainPassword) < 1) {
-                $error = "Invalid password length";
+                $error = "Die Passwortänge ist ungültig (0 < Länge < 4097).";
             }
             //create hashed password
             $password = $this->get('security.password_encoder')
@@ -132,12 +139,19 @@ class UserController extends Controller
             //set password
             $new_user->setPassword($password);
             $new_user->setIsActive(true);
+            //do not save user if there was an error
+            //render create template with error message
+            if($error){
+                return $this->render('user/create.html.twig', array(
+                    "message" => $error,
+                ));
+            }
             //save user to db
             $em->persist($new_user);
             $em->flush();
             //redirect to show user with new user
             $user_id = $new_user->getId();
-            $this->redirectToRoute("users_show", array("id" => $user_id));
+            return $this->redirectToRoute("users_show", array("id" => $user_id));
         }
         return $this->createNotFoundException("Not Found");
     }
@@ -153,9 +167,19 @@ class UserController extends Controller
         if(!$user){
             $error = "cannot delete user, user does not exist";
         }
-        $em->remove($user);
-        $em->flush();
-        $this->redirectToRoute("users_index");
+        // if method is get: render template "do you really want to delete?"
+        if($request->getMethod() === "GET"){
+            return $this->render('user/delete.html.twig', array(
+                "user" => $user,
+            ));
+            //if method is post (or anything else really) -> delete user
+        }else{
+            $em->remove($user);
+            $em->flush();
+            //TODO: eventuell template mit "user wurde erfolgreich gelöscht" dazwischen setzen
+            return $this->redirectToRoute("users_index");
+        }
+
     }
 
     private function roleExists($role)
